+++ models/db.js
@@ -1,0 +1,16 @@
+// import User from "./userModel.js"
+import mongoose from "mongoose"
+
+
+// CONNECT TO DATABASE
+mongoose
+  .connect(process.env.MONGODB_URI, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+  })
+  .then(() => {
+    console.log("Connected to MongoDB");
+  })
+  .catch((err) => {
+    console.error("Error connecting to MongoDB:", err);
+  });
