+++ package.json
@@ -3,15 +3,20 @@
   "version": "1.0.0",
   "description": "",
   "main": "app.js",
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "nodemon app.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "axios": "^1.6.8",
+    "dotenv": "^16.4.5",
+    "mathjs": "^13.1.1",
     "moment-timezone": "^0.5.45",
+    "mongoose": "^8.6.3",
+    "nodemon": "^3.1.0",
     "telegraf": "^4.16.3"
   }
 }
