+++ startMessage/menuKeyboard.js
@@ -1,0 +1,10 @@
+import {  Markup } from "telegraf";
+
+const menuKeyboard = Markup.keyboard([
+    ["ADMIN"],
+    ["INVITE", "BALANCE"],
+    ["HELP & SUPPORT"],
+    ["ADMIN PANEL"],
+  ]).resize();
+
+  export  {menuKeyboard}
