+++ startMessage/messages.js
@@ -1,0 +1,58 @@
+import { menuKeyboard } from "./menuKeyboard.js";
+
+const startMessages = [
+    {
+      message:
+        "Ｈｅｌｌｏ ${username}! Ｙｏｕ ｈａｖｅ ｔｏ ｊｏｉｎ ｔｈｅｓｅ ｃｈａｎｎｅｌｓ ｔｏ ｕｓｅ ｔｈｉｓ ｂｏｔ\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+  
+    {
+      message:
+        "ℍ𝕖𝕝𝕝𝕠 ${username}! 𝕐𝕠𝕦 𝕙𝕒𝕧𝕖 𝕥𝕠 𝕛𝕠𝕚𝕟 𝕥𝕙𝕖𝕤𝕖 𝕔𝕙𝕒𝕟𝕟𝕖𝕝𝕤 𝕥𝕠 𝕦𝕤𝕖 𝕥𝕙𝕚𝕤 𝕓𝕠𝕥\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+    {
+      message:
+        "H̥ͦe̥ͦl̥ͦl̥ͦo̥ͦ ${username}!̥ͦ ̥ͦY̥ͦo̥ͦu̥ͦ ̥ͦh̥ͦḁͦv̥ͦe̥ͦ ̥ͦt̥ͦo̥ͦ ̥ͦj̥ͦo̥ͦi̥ͦn̥ͦ ̥ͦt̥ͦh̥ͦe̥ͦs̥ͦe̥ͦ ̥ͦc̥ͦh̥ͦḁͦn̥ͦn̥ͦe̥ͦl̥ͦs̥ͦ ̥ͦt̥ͦo̥ͦ ̥ͦu̥ͦs̥ͦe̥ͦ ̥ͦt̥ͦh̥ͦi̥ͦs̥ͦ ̥ͦb̥ͦo̥ͦt̥ͦ\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+    {
+      message:
+        "нєℓℓσ ${username}! уσυ нανє тσ נσιи тнєѕє ¢нαииєℓѕ тσ υѕє тнιѕ вσт\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+    {
+      message:
+        "ℌ𝔢𝔩𝔩𝔬 ${username}! 𝔜𝔬𝔲 𝔥𝔞𝔳𝔢 𝔱𝔬 𝔧𝔬𝔦𝔫 𝔱𝔥𝔢𝔰𝔢 𝔠𝔥𝔞𝔫𝔫𝔢𝔩𝔰 𝔱𝔬 𝔲𝔰𝔢 𝔱𝔥𝔦𝔰 𝔟𝔬𝔱\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+  
+      keyboard: menuKeyboard,
+    },
+    {
+      message:
+        "ʜᴇʟʟᴏ ${username}! ʏᴏᴜ ʜᴀᴠᴇ ᴛᴏ ᴊᴏɪɴ ᴛʜᴇsᴇ ᴄʜᴀɴɴᴇʟs ᴛᴏ ᴜsᴇ ᴛʜɪs ʙᴏᴛ\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+    {
+      message:
+        "H͎e͎l͎l͎o͎ ${username}!͎ ͎Y͎o͎u͎ ͎h͎a͎v͎e͎ ͎t͎o͎ ͎j͎o͎i͎n͎ ͎t͎h͎e͎s͎e͎ ͎c͎h͎a͎n͎n͎e͎l͎s͎ ͎t͎o͎ ͎u͎s͎e͎ ͎t͎h͎i͎s͎ ͎b͎o͎t͎\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+    {
+      message:
+        "ⱧɆⱠⱠØ ${username}! ɎØɄ Ⱨ₳VɆ ₮Ø JØł₦ ₮ⱧɆ₴Ɇ ₵Ⱨ₳₦₦ɆⱠ₴ ₮Ø Ʉ₴Ɇ ₮Ⱨł₴ ฿Ø₮\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+    {
+      message:
+        "ɦɛʟʟօ ${username}! ʏօʊ ɦǟʋɛ ȶօ ʝօɨռ ȶɦɛֆɛ ƈɦǟռռɛʟֆ ȶօ ʊֆɛ ȶɦɨֆ ɮօȶ\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+    {
+      message:
+        "H͓̽e͓̽l͓̽l͓̽o͓̽ ${username!͓̽ ͓̽Y͓̽o͓̽u͓̽ ͓̽h͓̽a͓̽v͓̽e͓̽ ͓̽t͓̽o͓̽ ͓̽j͓̽o͓̽i͓̽n͓̽ ͓̽t͓̽h͓̽e͓̽s͓̽e͓̽ ͓̽c͓̽h͓̽a͓̽n͓̽n͓̽e͓̽l͓̽s͓̽ ͓̽t͓̽o͓̽ ͓̽u͓̽s͓̽e͓̽ ͓̽t͓̽h͓̽i͓̽s͓̽ ͓̽b͓̽o͓̽t͓̽\n\n<a href='https://t.me/cheemda_bot_testing_channel'>@cheemda_bot_testing_channel</a>",
+      keyboard: menuKeyboard,
+    },
+  ];
+
+  export{ startMessages}
